Removed precarious circular import with WidgyPage
There is a new get_widgypage_model method that mirrors
django.contrib.auth.get_user_model.
