+++ widgy/contrib/widgy_mezzanine/models.py
@@ -1,3 +1,5 @@
+from django.utils.translation import ugettext_lazy as _
+from django.conf import settings
 from django.core import urlresolvers
 
 from widgy.models.links import LinkableMixin
@@ -23,3 +25,22 @@ class WidgyPageMixin(LinkableMixin):
                 'form_node_pk': form.node.pk,
                 'root_node_pk': widgy['root_node'].pk,
             })
+
+
+# In Django 1.5+, we could use swappable = True
+if getattr(settings, 'WIDGY_MEZZANINE_PAGE_MODEL', None) is None:
+    from mezzanine.pages.models import Page
+    from widgy.db.fields import VersionedWidgyField
+
+    class WidgyPage(WidgyPageMixin, Page):
+        root_node = VersionedWidgyField(
+            site=settings.WIDGY_MEZZANINE_SITE,
+            to=getattr(settings, 'WIDGY_MEZZANINE_VERSIONTRACKER', None),
+            verbose_name=_('widgy content'),
+            root_choices=(
+                'page_builder.Layout',
+            ))
+
+        class Meta:
+            verbose_name = _('widgy page')
+            verbose_name_plural = _('widgy pages')
